requirements.txt
Django==4.1.3
djangorestframework==3.14.0
ipython==8.7.0
coverage==6.5.0